[22-05-2024] Half a page, not bad (GNN)
More seriously, findind reference and thinking about the way of telling
things is so time consuming.

chapters/machine_learning
  - Complete section about GNN
  - Maybe add a figure to illustrate ? Probably with CNN figure

main/
  - Add a library for \Box (it seems to not be ... out of the box !)

More general feeling: This chapter is gonna be short. Depend if I decide
to take time to explain the problematics of gradient descent, local
minima etc... I mean, those are thing I studied and fought againt but
does it have its place in a physic thesis ? Goal is not so much to
present the history of my struggles but more to present the end work.

I'm feeling personaly ok to have a ~5 pages chapter just introducing NN
and such. It's not very "scientific" in a way, it looks more like a
lesson or something but keep in mind the public is not expert. Maybe
such a chapter can be refreshing ?

To discuss w/ Fred and Benoit.